refactor: Merge handleAuth on oauth-callback and auth effect on useAuth
