+++ apps/main-landing/src/app/creators/page.tsx
@@ -1,14 +1,18 @@
 'use client';
 import { ScrollArea } from '@idriss-xyz/ui/scroll-area';
+import { useSearchParams } from 'next/navigation';
 
 import Content from './content';
 import { OAuthCallbackHandler } from './components/oauth-callback-handler';
 
 // ts-unused-exports:disable-next-line
 export default function Landing() {
+  const searchParameters = useSearchParams();
+  const customToken = searchParameters.get('token');
+
   return (
     <div className="relative flex h-screen">
-      <OAuthCallbackHandler />
+      <OAuthCallbackHandler authToken={customToken} />
       <ScrollArea
         type="always"
         customScrollEventName="creatorsLandingPageScroll"
